+++ sdk/storage/azure-storage-files-shares/inc/azure/storage/files/shares/share_file_client.hpp
@@ -10,7 +10,6 @@
 #include <azure/core/response.hpp>
 #include <azure/storage/common/storage_credential.hpp>
 
-#include "azure/storage/files/shares/protocol/share_rest_client.hpp"
 #include "azure/storage/files/shares/share_client.hpp"
 #include "azure/storage/files/shares/share_directory_client.hpp"
 #include "azure/storage/files/shares/share_options.hpp"